Add local PHP_CodeSniffer, installer, and standard
This package should have code that follows PSR-2. Installing PHP_CodeSniffer allows us to check for that.

We also state that this package should work for PHP 5.3 and above, so the PHPCompatibility package is a PHPCS standard that allows us to check that we're not using syntax from above PHP 5.3.

The phpcodesniffer-composer-installer package is a Composer plugin that allows for automatic registering of code standards outside of the PHPCS core.